\docType{data}
\name{MethylSig}
\alias{MethylSig}
\title{MethylResolver leukocyte reference signature matrix}
\usage{
MethylSig
}
\description{
Beta' values of 419 optimal CpGs for deconvolving methylation mixtures which contain leukocytes. The leukocytes in the signature include: Monocytes, Dendritic cells, Macrophages, Neutrophils, Eosinophils, Regulatory T cells, Naive T cells, Memory T cells, CD8 T cells, Natural Killer cells, and B cells.
}
\keyword{datasets}